test: Disable NPCX system for Google Quincy Zephyr tests
This commit updates the run_device_tests.py script to disable
CONFIG_CROS_SYSTEM_NPCX when building Zephyr tests for the google_quincy
board. This is necessary because the google_quincy board does not
require the NPCX crps system in this context, and explicitly disabling it
prevents potential build issues.

The upstream tests on the `google_quincy` board were failing to build
with an "bbram.h file not found" error. This was happening because
`CONFIG_CROS_SYSTEM_NPCX` was enabled by default (as `google_quincy` is
an `SOC_FAMILY_NPCX` platform), which pulls in `cros_system_npcx.c` and
its dependency on `bbram.h`.

BUG=b:419617755, b:419704767, b:419630887
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     zephyr_kernel_poll --zephyr
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     zephyr_drivers_entropy --zephyr
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     zephyr_cpp_newlib

Change-Id: I34604237ebd8a65f0b88628e602912159dc222d3