zephyr: update SMF users for API change
The SMF API changed to require a return value for the "run" action.
Update all SMF users to return the correct status.

The PDC subsystem and drivers are flat state machines (no parent states)
so the return value for "run" actions is ignored, but SMF_EVENT_HANDLED
is the preferred return value.

The AP power sequence driver defines parent states, so it always returns
SMF_EVENT_PROPAGATE because the driver never called smf_set_handled().

This should be merged with
https://github.com/zephyrproject-rtos/zephyr/pull/83854

BUG=none
TEST=./twister ivc -T zephyr/test/ap_power -T zephyr/test/pdc
TEST=Boot image on brox, verify PDC functionality

Cq-Depend: chromium:6653269
Change-Id: I0080cebf7e5f5c3cc350d38c8314916b0880920f