ocelot: Enable fans
Configure fan support for the ITE and NPCX AICs. Maximum RPM is based on
label on the fan, and the minimum RPM was determined experimentally.
This controls the fan connected to the J1D4 header (the one adjacent to
the USB-A port). The other fan header, J1A2, is not under EC control and
will spin at full speed if connected. Note: both fans require the S5
rails to be enabled.

BUG=b:424228054
TEST=zmake build ocelotrvp*
TEST=Use faninfo/fanset on ITE and NPCX AICs to control fan speed.
TEST=Monitor PWM and TACH signals on a scope

Change-Id: I7f84d695d65b6a0c3b8907076558f3efb5574404