driver/cec/it83xx: reduce F/R of "turn on TV, turn on chromebox"
For the chip go to deep doze when the system is at S3/S5/G3, the
clock is gating that the CEC will miss the message from the device.
Force the chip not enter dsleep when CEC is enable to reduce the
CEC missing the message.

BUG=b:411181188, b:406416769
TEST=test "turn on TV, turn on chromebox" 5 times. Only got 1 fail.
LOW_COVERAGE_REASON=b/414662955 tracking ztest.

Change-Id: I32d91b1188fa8df3fcd0e4dad5017e908bb1c87f