pdc: tps6699x: Handle irqs separately from SMF
When PDC_IRQ_EVENT is handled by st_idle_run, it's possible for
st_wait_task_run to query the data register continuously and starve the
TI PDC from executing. To avoid this, refactor the IRQ handling to
remove it from SMF and run on every loop of tps_thread.

BUG=b:423899597
TEST=./twister -cv --coverage -T zephyr/test/pdc/

Change-Id: Ic0efe52f856a4f8b7e839eed78390ae6647b3026