skywalker: enable DP altmode event as wake-source
Enable the MKBP_EVENT_DP_ALT_MODE_ENTERED as a wake source for the
skywalker platform. This allows the system to wake from S3 sleep when
a DP device is connected, triggering the altmode event.

This change modifies the `ec-mkbp-event-wakeup-mask` in the device
tree to include the DP altmode event in the wakeup mask.

BUG=b:424707361
TEST=(1) In S3, plug-in DP plug to DP dongle at C0/C1
TEST=(2) With this change, wakes up
TEST=(3) Without this change, doesn't wake up

Change-Id: Ic75a873f582a047a76bf98de6767beaa138250f8