pdc: Update RTK PDC console updater to take a raw I2C argument
Currently the PDC console updater targets a particular PDC chip by USB-C
port number. This assumes that the current EC image has matching
devicetree and/or runtime port configuration, but this is often not the
case for bringups or experimental PDC images.

Update the console command to accept a port number or a direct I2C bus
name and address. When an I2C bus and address are passed, the updater
will communicate directly to that I2C target.

Use the platform/pdc/scripts/pdc_console_fwup.py script to interact with
this feature.

BUG=b:417491139
TEST=zmake build
TEST=Reflash some PDCs over the console using port numbers and I2C addrs

Cq-Depend: chromium:6649326
Change-Id: I91ed633cff5945719a8c13f08dfae29bab2b586c