motion_sense_fifo: trust timestamp from handler when spreading
After sending data to host, reset the timestamp used for spreading.
It has been proven to be reliable, even when the EC is slow to process
sensor fifo.
Remove `read_ts` field as well, unused.

Fixes: 1c316fc5d23 ("motion_sense_fifo: Spread using the expected period")
BUG=b:413177152,b:417147201
TEST=Unit test, test on pujjoniru:
- cheets_CTS_T.internal.arm.CtsSensor
- cheets_CTS_T.internal.arm.CtsSensorTestCases.ctshardware
- tast run --repeats=30 <IP> hardware.SensorIioserviceHard
- Check on skywalker.
Noticed in dmesg messages like:
```
[ 3430.589692] cros-ec-sensorhub cros-ec-sensorhub.2.auto: 100
timestamps in the future, 36585002ns shaved on average
[ 3431.719494] cros-ec-sensorhub cros-ec-sensorhub.2.auto: 100
timestamps in the future, 44303476ns shaved on average
[ 3432.921898] cros-ec-sensorhub cros-ec-sensorhub.2.auto: 100
timestamps in the future, 52219357ns shaved on average
[ 3433.994990] cros-ec-sensorhub cros-ec-sensorhub.2.auto: 100
timestamps in the future, 59620044ns shaved on average
```

Change-Id: I8db5820fff10587e58ff0682dfffa71a0ddbba48