mutex: Add common recursive mutex support
This change adds support for a common recursive implementation of
mutex. Nested or recurive locks are supported within the same class.

The exiting libc/recurisve_mutex implmentation is leveraged by
symlinking the files to common/.

Since non-recursive mutexes are prone to deadlock that result in
watchdogs, this implementation is on by default.

BUG=b:425486970,b:418077172
TEST=Nested locks do not cause watchdog, tested on ITE and NPCX

Change-Id: Ifb01e02c7abe36f74655e4f4263ad977268aac72