mutex: Add common recursive mutex support
This change adds support for a common recursive implementation of
mutex. Nested/recursive locks are supported within the same task.

The exiting libc/recurisve_mutex implementation is leveraged by
symlinking the files libc/recursive_mutex files to common/.

Since non-recursive mutexes are prone to deadlock that result in
watchdogs, this implementation is turned on by default.

mutex_trylock and mutex_recursive tests are enabled for host. Minor
changes made to the tests to make them compatible with host emulator.

BUG=b:425486970,b:418077172
TEST=Nested locks do not cause watchdog, tested on ITE and NPCX
TEST=firmware.EcStress passes on Bookem and Bugzzy
TEST=Pass mutex, mutex_trylock and mutex_recursive host tests

Change-Id: Ifb01e02c7abe36f74655e4f4263ad977268aac72