pdc: Always evaluate PDOs after new Src Caps
Currently the PDC power manager will not evaluate PDOs if there is
already a sink path enabled. If a partner device sends initial Src
Caps followed by a second set of Src Caps, the EC may request a PDO from
the first set, then update it's internal copy of PDOs based on the
second set. But, if it does not re-evaluate the PDOs it will forward the
new PDO at the same index as the intially selcted PDO to the power
manager. If the second set of PDOs offer more power, than the device may
start to draw current based on the second set of PDOs and exceed the
existing PD contract.

BUG=b:424219903
TEST=TEST.PD.PS.SNK.1 passing with additional update to
READ_POWER_LEVEL.

Change-Id: Idf4911f64cb6d30d25020449f6e5396a9b599ba1