WIP: fatcat: Auto-set MODEL_ID tag in CBI
Automatically set the MODEL_ID tag in CBI iff it hasn't been set.

The value assigned to MODEL_ID is determined from the associated GPIOs
on the RVP. This is essentially the same value currently encoded in the
ChromeOS board id for the RVP.

Note: BOM_ID is missing in the original encoding.

the ChromeOS board ID should be a simple monotonically increasing number
that represents board revisions. It should not be overloaded with
additional bit fields.

TODO: Do we need the RVP BOARD_ID bits in MODEL_ID or are they
appropriate for the ChromeOS board ID?

Once we have the MODEL_ID implemented on the EC side, we'll need to
switch the AP firmware to use it. When that's complete, we can revert
the board ID to a simple number.

BUG=b:427302886

Change-Id: I85377709cadcd787db9d95b21733d9468186db3d