kanix: keyboard: do not set anything on cbi error
If cros_cbi_get_fw_config then the value in val is whatever happen to be
on the stack, bail out from the function after printing the error rather
than do something depending on the value, which could lead to
intermittent failures hard to troubleshoot.

Found running valgrind on rex.kanix.keyboard.

BUG=none
TEST=valgrind ./zephyr/zephyr.exe on rex.kanix.keyboard
TEST=cq dry run

Change-Id: I7db879f1dd249144b0695a488909ffad2c8b873d