ocelot: Use only PRIM_VR_EN to enable 3.3V and 5V
In revised RVP variants of LP5, only one GPIO signal is used for the
generation of 3.3V and 5V generation and the delay required for it, is
taken care by LDO. But, for Microchip based RVP, the old RVP variant
where 3.3V and 5V GPIOs are controlled individually are still under
use. Hence, 5V GPIO signal for power sequencing for Microchip is still
under usage and is not removed. But, usage of GPIO for 5V for the power
sequencing using ITE & Nuvoton based ECs are removed.

BUG=b:430093425
TEST=Boot ocelotrvp variants using Npcx/ITE/Mchp ECs

Change-Id: I932a511a93c6eb19033814f6dbc2f9b3e0815076
Signed-off-by: Poornima Tom <poornima.tom@intel.com>