pdc: Enable FRS until Snk Caps are evaluated
Currently, FRS is disabled by default until partner Snk Caps are
evaulated and the DPM confirms that the port is able to support the
partner's current requirements. That delays enabling FRS enough to fail
FRS compliance tests which trigger an FRS ~1.5 seconds after initial
connection.

This updates the PDC power manager and DPM to request 3A for PD source
partners to enable FRS while the device evaluates partner Snk Caps. If
FRS is not supported or 3A is not required, the DPM updates policy
accordingly.

BUG=b:405001062
TEST=./twister -T ./zephyr/test/pdc, TEST.PD.FRS.SNK*

Change-Id: Iff52f353224edbd95032360dab2e2f226c762b86