util/flash_fp_mcu: Make gpioset hold the GPIO value
As mentioned in "gpioset -h", the value of the GPIO is not guaranteed to
persist after gpioset exits:

The state of a GPIO line controlled over the character device reverts to
default when the last process referencing the file descriptor
representing the device file exits.  This means that it's wrong to run
gpioset, have it exit and expect the line to continue being driven high
or low. It may happen if given pin is floating but it must be
interpreted as undefined behavior.

On some devices (brya, jubilant) the gpioset value appears to persist
after gpioset exits, but on other devices it does not (coachz, volteer).

BUG=b:358420799, b:381902143
TEST=coachz:
     flash_fp_mcu /opt/google/biod/fw/\
       dartmonkey_v2.0.2887-311310808-RO_v2.0.28102-64c3d5971-RW.bin

Change-Id: I8e396e6d5c2cad93749c4a6c58438cece5d6264f