zephyr: scripts: Ignore disabled nodes in named_gpios.py
The named_gpios.py script makes sure that GPIOs defined in DT nodes
match the flags used for that pin in the named-gpio node. However, it
considers every node in the devicetree, which can lead to false
positives where a GPIO might be declared in an unused/disabled node as a
default or alternate pin muxing.

Update the checker script to ignore nodes that have a status of
"disabled".

BUG=None
TEST=Resolves build error

Change-Id: I026bae0a5f7f039afb63f5c4ee30b04c30078bbe