driver/cec/bitbang: fix CEC follower ACK timing for IT8XXX2
Due to ISR latency in Zephyr, the CEC follower cannot release the GPIO from LOW to HIGH within the required 1.7 ms. This delay causes the TV to stop sending subsequent data and to retry until the retry limit is reached. As a result, the factory test board fails during CEC communication tests.

BUG=b:426004099
TEST=Dirks can pass with the factory test board

Change-Id: Ib1b4c0d11004f77c509f337e4128dec7b942a828