TEST_ONLY:re-sign-gsc: add epoch and major version args
You can use this to create DBG images. The filename is a required arg.
The epoch is required. The major version is optional. If it's not given,
the script will keep the major version from the original image.

The minor version is the same as the original image

BUG=none
TEST=manual

Create DBG image
./util/re-sign-gsc.sh $DEVID ti50.bin 1

gsctool -b shows the version is 1.26.210

Create Ti50 DBG "low_ver" image
./util/re-sign-gsc.sh $DEVID ti50.bin 0 20

gsctool -b shows the version is 0.20.210

Change-Id: I1b978249175af3596bcbc17efa22cdf3a311808d