pm: Disable all PM state in high throughput cases
When the motionsense task is running a "high" throughput configuration,
we should avoid letting the CPU suspend. This is normally avoided by
having 1 sensor in a polling mode since the wait timeout will be set to
a value matching the expected next sample, but when all sensors are
using interrupts, this assumption breaks.

Add a new function to system.h that is only valid in Zephyr based builds
(by not implementing it under common/ we guarantee that the build will
break if the function was used in legacy ECOS). When called, this
function will loop through all the power states and lock them.
Similarly, a "put" variant will loop through all the power states and
unlock them.

BUG=b:427883250
TEST=5x tast run $DUT_IP hardware.SensorIioserviceHard
TEST=[pending] Sensor CTS

Change-Id: I7fa71c7ba09858cc6eec8e6f15833fb7693dc6b5