pdc_rts54xx: Add battery status/capability reporting over vendor command.
This patch introduces support for battery status and capacity reporting via vendor commands to the RTS54xx pdc driver.
It adds a callback to the battery SOC change hook to update the capacity and status on any change.
It adds a function that will update the status and capacity on S0,S3 resume and also during system startup.

This is needed to comply with USB PD spec.

BUG=b:412291685
TEST=manual

Also some refactoring on the code.

Change-Id: Iec5ddc45f309ec92a7f0e7193a6cf44c462311d8
Signed-off-by: David D. <david.delgado@wdc.com>