fpsensor: Keep ECDH private key on FPMCU while generating Pairing Key
In the Fingerprint Login implementation, the host command responsible
for generating FPMCU's ECDH key pair is returning encrypted private
key. The key is provided as a parameter to host command that finishes
ECDH and returns encrypted Pairing Key.

Let's keep ECDH private key in the FPMCU and deallocate it when Pairing
Key generation process is finished.

BUG=b:392915518
TEST=./test/run_device_tests.py -b bloonchipper -f servo_micro \
     -c clang -t fpsensor_auth_crypto_stateful

Change-Id: Ia2ddd7d798d4150d57b36dd6c423f1c756826157