battery: allow status to be neither charging nor discharging
On ACPI platforms, the EC_BATT_FLAG_DISCHARGING and _CHARGING flags in
battery dynamic information directly control the reported charge and
discharge state reported to host firmware, but previously always
reported the battery was charging if external power was connected even
if the battery was not accepting any power.

This changes the behavior from being either charging or discharging and
never neither, to report neither charging nor discharging if in a state
where the battery is not sourcing or sinking any power.

BUG=b:378555338
TEST=On redrix, inspect /sys/class/power_supply/BAT0/status in various
     states:
      * Allow battery to charge to full: Not charging, or Full
	(depending on whether reported charge state actually reaches
	100%, due to kernel ACPI battery driver behavior).
      * ectool chargecontrol discharge: Discharging
      * ectool chargecontrol idle: Not charging
      * ectool chargecontrol normal: Charging (when battery is not full)
TEST=ChromeOS UI still shows expected states:
      * Charging icon in shelf when a charger is connected
      * No charging icon when no charger is connected
      * "Battery full" in quick settings when fully charged

Change-Id: I0f0af0125965f61a2affca0f1270805bb8fb52da