npcx: set PM options from Kconfig instead of board
PM on NPCX needs the custom policy which is implemented in application
code in the pm_policy_next_state() function, drop the PM enable setting
from the board so that the board definitions can be used with non EC
application. This is how boards are defined in Zephyr in general, many
tests can't quite handle power management for whatever reason.

No changes for current projects.

BUG=none
TEST=zmake compare-builds -a

Change-Id: Ibd18e7f5d6960a307a55a846b3de16570a42a838