tps6699x: Set ucsi cmd base on port_index_on_chip
If there are two single port chips on the device, then the driver
should sent ucsi command with index 1.

If there is one dual-port chip on the device, then the driver
should sent ucsi command with index 1 or 2 respectively.

We add a new property port_index_on_chip to enable configuring the
combination in dts.

BUG=b:425235885
TEST=(0) Get a device with two single port TPS board.
TEST=(1) zmake build skywalker
TEST=(2) ./util/flash_ec --board skywalker --zephyr
TEST=(3) `pdc status 1` in ec console.
TEST=(4) Without this change, ec console spams logs.
TEST=(5) With this change, outputs `Enable`.
TEST=(6) Get a device with one dual-port TPS board.
TEST=(7) zmake build uldernite
TEST=(8) ./util/flash_ec --board uldernite --zephyr
TEST=(9) Both ports use correct ucsi index.
TEST=(A) ./twister -ivc -T zephyr/test/pdc

Change-Id: Icc8cc16059ac3f14b8f11e7a0c40d75c13b120df