fpsensor: Do not generate pairing key in host command buffer
The host command buffer is used to provide request to the host command
and collect response from it. The buffer is defined in the host commands
core implementation and it's shared with SPI/UART/USB often in
conjuntion with DMA.

For security reasons, we should never use it as a temporary memory for
processing sensitive data in host commands.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands run-fpsensor_auth_commands_otp \
     run-fpsensor_auth_crypto_stateful
TEST=./test/run_device_tests.py -b bloonchipper -f servo_micro \
     -c clang -t fpsensor_auth_crypto_stateful

Change-Id: I37d50ec8d0637c8359a4254ef37d6da001cea5b1