npcx_monitor, rtk_flame: decouple the flash loader from the shim directory
The flash support binaries for NPCX and RTK are currently under the shim
directory, which only gets included if PLATFORM_EC is set. These two are
useful even if the build does not include PLATFORM_EC, such as building
upstream tests against the EC board definitions.

Move the two scripts outside of shim and always build them when the
respective platform is selected.

BUG=none
TEST=zmake build rex && ./flash_ec --zephyr
TEST=./twister -vvv -p it8xxx2/it82002aw -p npcx9/npcx9m7f -p
realtek/rts5912 --toolchain=coreboot-sdk # verify that the loader
binaries are built

Change-Id: I5e62a16ac5852363e392fae5d9250a26fa2bca48