pdc: Verify port number on public functions, add more unit tests
Many of the public-facing pdc_power_mgmt API functions do no verify that
the caller-provided port parameter is in range, which can lead to
crashes when used to index into the `pdc_data` array.

Add `is_pdc_port_valid()` checks to these functions. Skip over static
helper methods taking a port parameter if the callers are already
checking the port number.

Update the documentation for these functions in the header file.

Update the unit tests for the modified functions. Some simply need an
additional check for an out-of-range port, and some had no testing at
all, in which case a basic test is added.

BUG=None
TEST=./twister -T zephyr/test/pdc

Change-Id: I77bb44415c1e0a2d285bdf640e53148121568aa4