rtk_flame: drop the RTKToolChain.cmake file
Drop the RTKToolChain.cmake file from the rtk_flame module, not sure
what was the intend of this but it seems to have no effect on the
Realtek loader build and breaks normal builds when c++ is enabled by
overriding c++_library with a value that does not work with
coreboot-sdk.

Dropping it does not seem to cause any other issues.

BUG=none
TEST=zmake build minimal-realtek && ../dagwood/flash_rtk.py -b minimal-realtek
TEST=./twister -vvv -p it8xxx2/it82002aw -p npcx9/npcx9m7f -p realtek/rts5912 --toolchain=coreboot-sdk

Change-Id: I38550dd20bb1b6bd6d35e6628a0b2e0249ca6716