zephyr: shim: Don't allow CONFIG_TEMP_SENSOR_POWER and power-good-pin
Some temperature sensors are on a different power rail and are only
valid when that rail is up. The temp sensor drivers allow checking a
GPIO to sense if the rail is up before taking a reading. Both the legacy
drivers and the Zephyr shim have mechanisms for specifying that GPIO.

For Zephyr boards, the correct configuration when power rail sensing is
required is to define the `power-good-pin` devicetree property in the
`cros-ec,temp-sensors` child nodes. Tagging a named-gpio with the
`GPIO_TEMP_SENSOR_POWER` enum is not necessary, and
`CONFIG_TEMP_SENSOR_POWER` should not be enabled.

Add a build assertion to ensure that power-good-pin and
CONFIG_TEMP_SENSOR_POWER are not used together.

In a future change, it should be possible to deprecate
CONFIG_TEMP_SENSOR_POWER entirely and migrate the few boards using it to
using power-good-pin instead.

BUG=None
TEST=zmake build -a
TEST=./twister

Change-Id: Ic77d4eecf36409e4d823ecd35e7c7804ef0fb3fa