pdc_rts54xx: Add battery status/capability over vendor command.
This patch introduces support for battery status and capacity reporting
via vendor commands to the RTS54xx pdc driver. It adds a callback to the
battery SOC change hook to update battery status on any battery change.
It adds a function that will update the status and capacity on src or
snk attach.

This is needed to comply with USB PD spec.

BUG=b:412291685
TEST=manual

Change-Id: Iec5ddc45f309ec92a7f0e7193a6cf44c462311d8