pdc: Make SET_PDR API less error-prone
The SET_PDR UCSI command controls the PDC's power direction and includes
a series of bits that encode the preferred power role to set the PDC to
and whether external power role swaps are accepted.

The PDC power management state machine has been haphazardly updating
these bits in a recycled buffer and issuing the SET_PDR command. There
is no provision to clear the buffer after the command is processed, and
there is no way to query the current PDR bits from the PDC. This makes
our current approach to using the SET_PDR command buggy, as seen in
b/432292811.

To remedy this, take away PDC power management's direct access to the
SET_PDR bits and instead allow it to choose one of four legal power role
modes: sink or source, with or without external swaps allowed. Each
mode is represented by an enum value that is passed to the set_pdr
method in the PDC API. The PDC API will translate the enum value into an
appropriate UCSI command. This ensures that all the SET_PDR bits are set
correctly and consistently.

Update the unit tests with the new API. Aside from bug fixes in the two
previous CLs, no functional changes were needed to the UTs.

BUG=b:435495591
TEST=./twister -T zephyr/test/pdc

Change-Id: Ie95b6f0d8cf54ab53675db3839fd9a1465801efe