fpsensor: Keep the session key for further use
In the Fingerprint Login approach, the session key was only used to
securely transport nonce context, aka user_id in classic approach.

In the new approach, the session key is also used to sign messages sent
between FPMCU and Trusted Application, so let's keep the session key for
further use.

This CL also renames the host command and the flag it sets to better
reflect its purpose.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands

Change-Id: I4742eb00fc52d78025ac5e1f6f0777cf944be0c0