fatcatrvp-npcx: Remove pull-up from WP signal
The EC add-in cards normally have a 100K resistor in series with the
MECC C_EC_WP_ODL signal which means we'll never see this pin go low with
an internal pull-up enabled. The original nuvoton EC AIC does not have a
resistor, so we didn't see this issue. Remove the pull-up anyway to
match the latest design.

BUG=b:418282403
TEST="gpioget wp_odl" toggles with GSC wp change

Change-Id: I8cc9fd48c1ea77568ef74322dbfb01f33e624fe4