usbpd: Add an overridable port number check function
Standardize the criteria for determining whether a USB-C port number is
valid (i.e. port is active and able to perform PD operations).

Create an overridable function board_pd_port_num_is_valid(). The default
implementation simply performs a range check (port must be >= 0 and less
than board_get_usb_pd_port_count().

The PDC subsystem overrides this function and adds an additional check
to ensure the target port is not in a disabled state --which happens
when the driver failed to initialize. (A PDC port is also disabled in
variable port count systems when not used, but this is already reflected
by the pdc_power_mgmt_get_usb_pd_port_count() count)

Use this function in the PD host commands and PDC console so we are
performing a consistent check.

BUG=None
TEST=Build various zephyr and legacy boards
TEST=./twister -T zephyr/test/pdc -T zephyr/test/drivers
TEST=CQ

Change-Id: I60318b5348a803aa2db7ec197ae85417335daada