col_gpio_drive: Enforce exact match for kso-ignore-mask
The previous validation for the ITE keyboard controller's
`kso-ignore-mask` was incomplete, only ensuring that `col-gpio`
nodes had a corresponding bit set in the mask.

This change enforces a strict 1:1 correspondence between the
mask and the defined `col-gpio` nodes.

BUG=b:433813612, b:435516562
TEST=Builds successfully. Manually verified that build fails
with the new error message if kso-ignore-mask does not exactly
match the col-gpio nodes.

Change-Id: I901343e446703b465bdb064bdd9cd132f7bc0e2e