lite-ish: Add lite ISH firmware
The lite ISH firmware does not have any sensor stacks, all the
sensor related I/O and interrupts are disabled. For the Trulo
devices which have the ISH shim loader stiched in the AP firmware,
this ISH firmware needs to be loaded into ISH after booting to OS
to allow the device to suspend normally.

BUG=b:439995676
TEST=zmake build lite-ish

Change-Id: I9c430bebb6017caa04df7234de381116e49279b7