system: add a new host command to enter bootloader
Add a new command to enter bootloader mode.

Such command can be needed for applications, if the bootloader is needed
to update firmware, but there is no additional pin to force EC to stay
in the bootloader after reboot.

Not all security consequences have been handled with this commit.
b/460674359 keeps track of this effort.

BUG=b:432672112
TEST=Add implementation of chip_enter_bootloader, send the command and
make sure chips enters bootloader.

Change-Id: I728409c2bc7cff0e68b415932e3a5d650d7d4a22