emul: rts545x: Align CCOM handing with TI driver, improve logging
This CL does not make functional changes. It adjusts the RTK PDC
emulator to store an `enum ccom_t` type instead of the raw UCSI CCOM
bitfield, which is how the TI driver handles this. This keeps the two
emulators consistent to avoid coding errors.

Improve the logging around SET_CCOM and SET_PDR so it is easier to debug
current PD settings applied on the emulator.

Note: the port numbers sent to the RTK emulator are all zero and they
are not worth printing in log messages.

BUG=b:435495591
TEST=./twister -T zephyr/test/pdc

Change-Id: I6bb1e9513b6465f4d6e4d84d64c5bb1efc52f607