uldrenite: Fix number of sensors for clamshell
There is only one accelerometer for clamshell, it should be the first
in motion_sensors array. Set the array to 3 sensors to accomodate
convertible. Adapt the sensors definition based on the CBI values.

Add a new Kconfig to enable dynamic on-body-sensor. When used, the
board is responsible for selecting the right index for the on-body
sensor. In the case of uldrenite, this means that in clamshell we'll
use the only sensor (lid_accel) as the on-body sensor; otherwise,
we'll use the alias.

BUG=b:435407060
TEST=Check on uldrenite: the sensor is presented as the first one.
     And activity sensor is the second sensor.

Change-Id: Ia954bac136d5b7ba7a88545e9e17bf480083b46c