Makefile.cq: add a target for bare board and coreboot-sdk
Add a build only test target using coreboot-sdk against the bare boards
for it82002aw, npcx9mf7f and rts5912.

This ensures that the base board definitions build cleanly against the
upstream tests, i.e. the EC application code does not introduce any
option or code that is enabled outside of the EC app and breaks the
build.

These three targets are the chips used in AIC boards, long term plan is
to use them to run actual tests on hardware.

At the moment this builds about 40 tests and takes about 1m to build on
a cloudtop instance.

BUG=none
TEST=time ./twister --outdir twister-out-coreboot-sdk -v -i --no-upload-cros-rdb --toolchain coreboot-sdk -p it8xxx2/it82002aw -p npcx9/npcx9m7f -p realtek/rts5912
TEST=make -f Makefile.cq test

Change-Id: If49762c1754063687c2e7dc890cdb45da046f65d