uldrenite: Fix invalid sensors on clamshell
We require all sensors to be in order and without gaps, since the base
accel was the second item in the list, clamshell versions of uldrenite
would have the motion_sensors[0] entry be invalid. The ordering in
motionsense.dtsi is swapped so the base accel is first.

Next, the sensor cound for clamshell is set to 1 so we avoid
initializing sensors that aren't there.

BUG=b:435407060
TEST=None yet

Change-Id: I2dce01fefbca6140342371a658fcd3cbd1ecd41e