uldrenite: Fix number of sensors for clamshell
There is only one accelerometer for clamshell, it should be the first
in motion_sensors array. Set the array to 3 sensors to accommodate
convertible. Adapt the sensors definition based on the CBI values.

Enable dynamic on-body sensor index and set it during sensor
initialization.

Note that this doesn't work on uldrenite with EC sensors because it
uses the legacy on-body detection which requires the sensor to be
statically allocated. Making the change to common/body_detection.c
would be too risky so it was left untouched. Also, Uldrenite with EC
sensors is not officially supported and is slotted to be removed soon.

BUG=b:435407060
TEST=Check on uldrenite: the sensor is presented as the first one.
     And activity sensor is the second sensor.

Cq-Depend: chromium:6919195
Change-Id: Ie03196df6b3bad22144bad3d0b540d7eb9599835