zephyr: LED: Add pwr_led_support property
There's feature that LED behavior will depend on whether power
LED is supported (battery LED will be overridden as power LED
in specific case if power LED is not supported), add the property
to support this feature.

BUG=none
TEST=verify on caboc that battery LED will work as power LED in
specific case if power LED is not supported (just for test, not
implemented in reality).
     twister -T drivers

Change-Id: Ie0665f124b99a030c506779e2822971fd4c2feea