host events: Add non-fatal panic event
When the ISH goes down it will restart but the device might not even
notice. We need a way for the ISH to trigger a non fatal crash report
on the AP. When the event is raised, it's expected that the AP will
collect the logs and submit the report.

BUG=b:439591195
TEST=CQ

Change-Id: I2457600d304e68a04393160e97bf22a5547b9be1