pdc: Make pdc_power_mgmt_request_power_swap() non-blocking
This function previously blocked the caller until the power role swap
completed. This has the potential to tie up other parts of the system
for a long time and lead to a WDT crash.

Modify this function to set policy flags instead, making the actual
power role swap asynchronous. This is also more consistent with the
TCPMv2 implementation.

BUG=b:465694662
TEST=zmake build hylia
TEST=./twister -T zephyr/test/pdc
TEST=No longer reproduce b/465694662#comment7 on hylia devices

Change-Id: Ia33c282a17451d43d13e3878cfd05365467438af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7319869
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>