fpsensor: Allow setting the context after the session is established
In the old Fingerprint Login approach, established session means that
the nonce context was set. The flag was checked in FP_CONTEXT command to
make sure that there is no downgrade to old protocol (the context was
set by other commands).

In the new Fingerprint Auth approach, established session means that TPM
seed is set (CL:6839792). The FP_CONTEXT command is used to set context
(provide user id), so there is no downgrade really.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands

Change-Id: Idea9bf5ba98fca976bfc5f521cb1b75134c03184