fpsensor: Do not invalidate existing session on context reset
In the Fingerprint Login approach, the nonce context had to be
invalidated on the context reset (because it's part of the context).

In the Fingerprint Auth approach, the Session Key is a per boot secret
used to decrypt the TPM Seed and sign/validate messages to/from Trusted
Application. It's not a part of the context, so let's not invalidate
it on the context reset.

Please note that the check if the session nonce is filled with zero was
removed from EC_CMD_FP_LOAD_PAIRING_KEY checks. This does not decrease
security, because it's not possible to establish session if the pairing
key was changed after session nonce was generated.

BUG=b:392915518
TEST=make run-fpsensor_auth_commands

Change-Id: Icc53ff776fb299c33c516dd12ddaf4db2b0f21ee