TCPMv2: Invalidate all suppliers in 0A contract
When the port as Sink enters into a contract at 0A, the Source expects
the Sink to sink no more than pSnkSusp = 25mW. Intended behavior in this
situation is to sink as close to 0 mW as possible by setting the current
limit to 0A. Even though it does not make sense for more than one
supplier on a given port to be active at the same time, charge manager
presently selects the highest-power supplier on a given port. To ensure
the 0A limit, the TCPM must limit any active suppliers to 0A.

Replace the existing calls to disable a specific subset of suppliers
with a single call to disable all possible suppliers on the port. This
may in some cases avoid a long-term current draw significantly above 0A
from a stale supplier. In practice, it also reduces the latency to reach
0A by 10s of ms.

This marginally improves performance on TEST.PD.PS.SRC.3#17, checking
pSnkSusp after entering a 0A contract, but it is not a complete
solution.

BUG=b:433547376
TEST=Observe 0A transition latency drop from ~95 ms to ~45 ms

Change-Id: Ia2a1ab66b45748d57e66088d1c7fa6148cb80e4f