gesture: Always define a valid LID_ACCEL
motion_sense.c requires LID_ACCEL to be defined when
PLATFORM_EC_GESTURE_DETECTION is set, even though it's only used if
CONFIG_ORIENTATION_SENSOR is enabled. This is caused by the use of
if(IS_ENABLED(...)) which will remove the use of LID_ACCEL at link
time. Until then, LID_ACCEL must be defined.

Default the LID_ACCEL to the invalid sensor ID if the lid_accel label
isn't set in devicetree in order to pass the compilation.

BUG=b:435407060
TEST=Build and test on uldrenite360

Change-Id: I4c708edda3d37dc125a3c26b72cd12bec44ece7b